Add HelloAndBye and Greeter classes
This demos how you can partially overwrite methods and leave the rest of
the methods untouched. In this example the `HelloAndBye.hello()` method
is patched and `HelloAndBy.bye()` is left untouched. I think this is
called a "stub" but I'm not entirely sure.

I find it especially interesting that the `hello` method is mocked after
the Greeter object was created. This is quite useful to mock methods
where they are called.